requirements.txt
flask==3.0.3
flask-cors==4.0.1
pdfplumber==0.11.0
python-docx==1.1.2
spacy==3.7.5
textstat==0.7.3
requests==2.32.3
python-dotenv==1.0.1
